requirements.txt
# Core Data Science Libraries
pandas==2.1.4
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2

# Financial Data APIs
yfinance==0.2.28
alpha-vantage==2.3.1
quandl==3.7.0
fredapi==0.5.1

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Visualization
plotly==5.17.0
dash==2.14.2
dash-bootstrap-components==1.5.0

# AWS
boto3==1.34.0
botocore==1.34.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.0
pytest==7.4.3
pytest-cov==4.1.0

# Portfolio Optimization
cvxpy==1.4.1
# pypfopt==1.5.5  # Commented out due to compatibility issues

# Time Series
arch==6.2.0
statsmodels==0.14.0

# Logging
loguru==0.7.2
